+++ build_runner/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bug fix: post process builders with `build_to: source` can only write to
   output packages.
 - Bug fix: reject incorrect builder config that has duplicate builder keys.
+- Bug fix: reject invalid builder factories in `build.yaml`.
 
 ## 2.15.2